move `test_name` and `test_support_html` to default testcase
